authenticate and authorize added

+++ src/services/contacts.js
@@ -3,22 +3,18 @@ import { ContactsCollection } from "../db/models/contact.js";
 import { calculatePaginationData } from "../utils/calculatePaginationData.js";
 
 
-export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_ORDER.ASC, sortBy='_id', filter={} }) => {
+export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_ORDER.ASC, sortBy='_id', filter={}, userID }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
 
-    const contactsQuery = ContactsCollection.find();
+    const contactsQuery = ContactsCollection.find({userId : userID});
 
     if (filter.contactType) {
         contactsQuery.where('contactType').equals(filter.contactType);
     };
 
-    // const contactsCount = await ContactsCollection.find().merge(contactsQuery).countDocuments();
-
-    // const  contacts = await  contactsQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder}).exec();
-
     const [contactsCount, contacts] = await Promise.all([
-        ContactsCollection.find().merge(contactsQuery).countDocuments(),
+        ContactsCollection.find({ userId: userID }).merge(contactsQuery).countDocuments(),
         contactsQuery.skip(skip).limit(limit).sort({ [sortBy]: sortOrder}).exec(),
     ]);
 
@@ -32,24 +28,28 @@ export const getAllContacts = async ({page = 1, perPage = 10, sortOrder = SORT_O
 };
 
 
-export const getContactById = async (contactID) => {
-    const contact = await ContactsCollection.findById(contactID);
+export const getContactById = async (contactID, userId) => {
+    const contact = await ContactsCollection.findOne({_id: contactID, userId: userId});
     return contact;
 };
 
-export const createContact = async (payload) => {
-    const contact = await ContactsCollection.create(payload);
+export const createContact = async (payload, userId) => {
+    const data = {
+        ...payload,
+        userId: userId,
+    };
+    const contact = await ContactsCollection.create(data);
     return contact;
 };
 
-export const deleteContact = async (contactID) => {
-    const contact = await ContactsCollection.findByIdAndDelete(contactID);
+export const deleteContact = async (contactID, userId) => {
+    const contact = await ContactsCollection.findByIdAndDelete({_id: contactID, userId: userId});
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (contactId, payload, userId, options = {}) => {
     const rawRes = await ContactsCollection.findOneAndUpdate(
-        { _id: contactId },
+        { _id: contactId, userId: userId },
         payload,
         {
             new: true,
